third_party/WebKit/LayoutTests/platform/win/virtual/outofblink-cors-http/http/tests/fetch/serviceworker-proxied/thorough/nocors-expected.txt

This is a testharness.js-based test.
PASS Startup 
FAIL Verify access control of fetch() in a Service Worker assert_equals: http://localhost:8000/serviceworker/resources/fetch-access-control.php?method=GET&headers=CUSTOM jsonpResult must be success expected "success" but got "error"
Harness: the test ran to completion.
